# Add check to speed up RP by 50 %