This extra check speeds up Rescue Prime on my DELL XPS 16GB RAM with 50
%.

Executing `cargo criterion hash` gives me:

Before this commit:
hash_rp64_256 (random)  time:   [31.142 us 31.200 us 31.292 us]

After this commit:
hash_rp64_256 (random)  time:   [17.937 us 17.951 us 17.966 us]

My best guess for why this significant speedup happens is that the
multiplication in `add` can sometimes be avoided.