@irakliyk 2022-06-03
By the way, the command I use to run benchmarks is `cargo bench hash` (not `cargo criterion hash`).