@Nashtare 2022-06-03
I'm also getting a nice improvement on `hash_rp64_256 (random)`. Down 39% when running `cargo bench hash` and down 43% when running `RUSTFLAGS="-C target-cpu=native" cargo bench hash`. I am on an x86 arch (running on Ubuntu 22.04).