@Sword-Smith 2022-06-04
> Interesting! On Apple M1 Pro I'm actually seeing this change making Rescue Prime run about 5% slower. I'll need to check on an x86 machine. But if this benefit shows up there, we could probably put a feature flag to add this modification only on x86 architecture.
> 
> One question: which version of rustc are you running this with? I'm using 1.61.0.

The observed speedup is confirmed on my setup with version `1.61.0` and also by running `cargo bench hash` -- `bench` gives me the same speedup as `criterion`.