@irakliyk 2022-06-06
I ran some more benchmarks, and here are my results (all results are for `hash_rp64_256 (random)` benchmark, showing before and after results):

* My original benchmark on M1 Pro (Mac OS): 8.6 us -> 9 us (about 5% degradation)
* x86 Windows (AMD Ryzen 9 5950X): 8.8 us -> 9.7 us (about 10% degradation)
* x86 Ubuntu (AMD EPYC 7R13): 28 us -> 16 us (about 43% improvement)

I'm not quite sure what to make of these numbers. It is obvious that the compiler is not doing its job correctly on AMD EPYC 7R13 Ubuntu because 28 us is way too slow to begin with. This extra bit of code helps it somehow, but it I'm not sure if this is related to the OS or to the machine architecture. It does seem like the issue may be caused by the branch predictor behavior, because even on AMD EPYC 7R13 the "cached" version of the benchmark yields 13 us result.

It is also possible that the benchmark itself is flawed in some way, and maybe in a context of a more complex computation,  the differences would not appear.