@Sword-Smith 2022-07-04
> @Sword-Smith - given #101 should we close this PR?

I think we should. I should compare both on my setup, but let's close. I'll reopen if I find a speedup.